Move code to pkg/
