+++ pkg/installing/uninstalling.go
