+++ pkg/paths/paths.go
