+++ pkg/paths/paths_test.go
