+++ pkg/teeth/rawmetadata.go
